Refactor ViewLoader to utilize ViewHelper class for improved template functionality and maintainability
